requirements.txt
PySide6>=6.5
pyserial
watchdog
openpyxl
pdfplumber
